Refactor UI modules into screens package with compatibility wrappers
