+++ docs/backlog.md
@@ -1,3 +1,4 @@
+- [x] UI-14 Arborescence écrans UI incrémentale: création de `game/ui/screens/`, déplacement progressif des modules de vue (`command_center`, `dashboard`, `facility`, `mission_board`, `research_lab`) avec wrappers de compatibilité temporaires dans `game/ui/`, et test de non-régression d'imports `tests/ui/test_screen_wrappers.py`. (completed May 22, 2026)
 - [x] UI-11 Feedback system modulaire: création de `game/ui/feedback/` (`toast_center.py`, `confirm_dialog.py`, `error_banner.py`), standardisation des messages `Action/Result/Impact/Next`, confirmations pour actions coûteuses/irréversibles (mission risquée, dépenses, load), et tests de rendu/dispatch `tests/ui/test_feedback_system.py`.
 - [x] UI-12 Fix room-action pulse helper import in graphical panels: `draw_action_button`/`draw_close_button` now resolve `pulse_from_elapsed` from `game/ui/theme/motion.py`, with a regression test to keep expanded-room buttons rendering safely. (completed May 22, 2026)
 TODO:
