+++ game/ui/screens/__init__.py
@@ -1,0 +1,5 @@
+"""UI screen-level modules.
+
+This package hosts per-screen builders and wrappers while migration away from
+flat ``game.ui`` modules proceeds incrementally.
+"""
